tests/wpt/tests/html/canvas/element/pixel-manipulation/2d.imageData.get.rounding.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.imageData.get.rounding</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.get.rounding</h1>
<p class="desc">Test the handling of non-integer source coordinates in getImageData().</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test the handling of non-integer source coordinates in getImageData().");
_addTest(function(canvas, ctx) {

  function testDimensions(sx, sy, sw, sh, width, height)
  {
      imageData = ctx.getImageData(sx, sy, sw, sh);
      _assert(imageData.width == width, "imageData.width == width");
      _assert(imageData.height == height, "imageData.height == height");
  }

  testDimensions(0, 0, 20, 10, 20, 10);

  testDimensions(.1, .2, 20, 10, 20, 10);
  testDimensions(.9, .8, 20, 10, 20, 10);

  testDimensions(0, 0, 20.9, 10.9, 20, 10);
  testDimensions(0, 0, 20.1, 10.1, 20, 10);

  testDimensions(-1, -1, 20, 10, 20, 10);

  testDimensions(-1.1, 0, 20, 10, 20, 10);
  testDimensions(-1.9,  0, 20, 10, 20, 10);

});
</script>
